+++ pkg/util/stack.go
@@ -1,0 +1,29 @@
+package util
+
+type Stack struct {
+	items []interface{}
+}
+
+func (s *Stack) Push(item interface{}) {
+	s.items = append(s.items, item)
+}
+
+func (s *Stack) Pop() interface{} {
+	if len(s.items) == 0 {
+		return nil
+	}
+	item := s.items[len(s.items)-1]
+	s.items = s.items[:len(s.items)-1]
+	return item
+}
+
+func (s *Stack) Peek() interface{} {
+	if len(s.items) == 0 {
+		return nil
+	}
+	return s.items[len(s.items)-1]
+}
+
+func (s *Stack) Size() int {
+	return len(s.items)
+}
